# public/openapi.yaml
openapi: 3.0.3
info:
  title: BEPRO
  description: BEPRO API Documentation
  version: 2.7.0
  contact:
    name: Hello
    email: hello@taikai.network
servers:
  - url: 'https://app.bepro.network/api'
    description: production api
  - url: 'https://aurelius.taikai.network/api'
    description: Aurelius TEST api
  - url: 'https://seneca.taikai.network/api'
    description: Seneca TEST api
  - url: 'https://diogenes.taikai.network/api'
    description: Diogenes TEST api
  - url: 'https://apollodorus.taikai.network/api'
    description: Apollodorus TEST api
components:
  schemas:
...
    NetworkDbId:
      type: string
      description: id of network in database
...
    PullRequestPayload:
      type: object
      properties:
        issueGithubID:
          type: string
        title:
          type: string
        description:
          type: string
        username:
          type: string
        branch:
          type: string
        networkName:
          type: string
    PullRequest:
      type: object
      properties:
        githubId:
          type: string
        issueId:
          type: integer
        githubLogin:
          type: string
        userRepo:
          type: string
        userBranch:
          type: string
        userAddress:
          type: string
        status:
          type: string
        contractId:
          type: string
        reviewers:
          type: array
          items:
            type: string
        network_id:
          $ref: '#/components/schemas/NetworkDbId'
...
paths:
...
  /pull-request/:
    get:
      description: retrieve one or more pull-requests from database
      parameters:
        - in: query
          name: login
          required: true
          schema:
            type: string
        - in: query
          name: issueId
          required: true
          schema:
            type: string
        - in: query
          name: networkName
          required: true
          schema:
            type: string
      responses:
        404:
          description: Invalid network, Issue not found
          content:
            application/json:
              schema:
                type: string
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items:
                      $ref: '#/components/schemas/PullRequest'
    post:
      description: create new pull-request in the database
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PullRequestPayload'
      responses:
        404:
          description: Invalid network, Issue not found
          content:
            application/json:
              schema:
                type: string
        422:
          description: Unprocessable entity
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  bountyId:
                    type: string
                  originRepo:
                    type: string
                  originBranch:
                    type: string
                  originCID:
                    type: string
                  userRepo:
                    type: string
                  userBranch:
                    type: string
                  cid:
                    type: string
  # DELETE operations can't have body, but we do. need to refactor so we can document it
  #    delete:
  #      description: Remove pull-request from database
  #      requestBody:
  #        content:
  #          application/json:
  #            schema:
  #              $ref: '#/components/schemas/PullRequestPayload'
  #      responses:
  #        500:
  #          description: Failed blockchain connection
  #          content:
  #            application/json:
  #              schema:
  #                type: string
  #        404:
  #          description: Invalid network, Issue not found, PR not found
  #          content:
  #            application/json:
  #              schema:
  #                type: string
  #        200:
  #          description: OK
  #          content:
  #            application/json:
  #              schema:
  #                type: string
...